analyzer/requirements.txt
tree-sitter-c>=0.23.0
tree-sitter>=0.23.0
matplotlib>=3.5.0
networkx>=2.8.0
numpy>=1.21.0
PyYAML>=6.0
